update install scripts

+++ ofGen/install.sh
@@ -1,4 +1,4 @@
-#!/bin/zsh
+#!/bin/sh
 cd "$(dirname "$0")"
 
 COLOR='\033[0;32m'
@@ -10,15 +10,18 @@ section() {
 }
 
 section "OpenFrameworks ofGen (tool to generate projects)"
-echo this will install a symlink in /usr/local/bin/ so ofGen can be called from any directory.
-echo if you need to uninstall in the future you can use the $ofw/ofGen/uninstall.sh script
-echo -------
-# if [[ -f /usr/local/bin/ofGen ]]; then
-# read -p "ofGen is already linked, overwrite? " -n 1 -r
-#     if [[ $REPLY =~ ^[Yy]$ ]]; then
-#     sudo rm /usr/local/bin/ofGen
-#     fi
-# fi
+echo This will install a symlink in /usr/local/bin/ so ofGen can be called from any directory.
+echo If you need to uninstall in the future you can use the $ofw/ofGen/uninstall.sh script
+echo It is recommended to install ofGen, skip if you already have it installed.
+echo You will be asked for user password
+# echo -------
+read -p "Proceed? (y/n) " -n 1 -r
+echo    # (optional) move to a new line
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
 # -sf parameters will overwrite old symlink if it is already installed
-sudo ln -sf "$PWD/ofGen" /usr/local/bin/ofGen
-echo All good!
+	sudo ln -sf "$PWD/ofGen" /usr/local/bin/ofGen
+	echo All good!
+else
+	echo OK, ofGen not installed
+fi
